refactor: seed npm script, changed file names, and methods to close db conneciton
