+++ semana19/labook/src/app/services/HashManager.ts
@@ -1,4 +1,7 @@
 import * as bcrypt from 'bcryptjs'
+import dotenv from 'dotenv'
+
+dotenv.config()
 
 export class HashManager {
   private static cost: number = Number(process.env.BCRYPT_COST)
