+++ semana19/labook/src/app/useCases/Post/CreatePost/CreatePostDTO.ts
@@ -1,3 +1,5 @@
+import { POST_TYPES } from '../../../types/Post'
+
 export interface ICreatePostRequestDTO {
   photo: any
   description: any
