+++ semana19/labook/src/database/DatabaseConnection.ts
@@ -1,10 +1,10 @@
 import dotenv from 'dotenv'
-import { default as knex, default as Knex } from 'knex'
+import { default as knex } from 'knex'
 
 dotenv.config()
 
 export class DatabaseConnection {
-  protected static connection: Knex = knex({
+  protected static connection: knex = knex({
     client: 'mysql',
     connection: {
       host: String(process.env.DB_HOST),
@@ -15,4 +15,18 @@ export class DatabaseConnection {
       multipleStatements: true,
     },
   })
+
+  protected static destroy() {
+    knex({
+      client: 'mysql',
+      connection: {
+        host: String(process.env.DB_HOST),
+        user: String(process.env.DB_USER),
+        password: String(process.env.DB_PASSWORD),
+        database: String(process.env.DB_SCHEMA),
+        port: 3306,
+        multipleStatements: true,
+      },
+    }).destroy()
+  }
 }
